New minimal CSS file.
This minimal CSS only contains the properties we want to modify from their original state.  
This means we still need most of the original CSS provided by site, this css file is meant as an "overwrite" over existing css.  
Especially for plugins like Stylish so we don't have to load the entire CSS file again with minor adjustments.