Deleting minimal.css as it serves no more purpose
The minimal.css was for the stylish plugin, while the full.css was for the work-around.  
As the work-around is no longer working, there's no need to distinguish between the 2.
Therefor I've decided to make the old "full.css" the new css you have to paste into stylish,
And decided to remove the old minimal.css as that is now the purpose of the new full.css.

I hope that made sense lol.